feat: move main to cmd/server to add cli
