+++ Dockerfile
@@ -78,7 +78,7 @@ CMD [ "/bin/bash",  "-c \"while sleep 1000; do :; done\"" ]
 
 FROM base AS builder
 
-RUN rm -rf build && go build -o build/${APP_NAME} cmd/main.go
+RUN rm -rf build && go build -o build/${APP_NAME} cmd/server/main.go
 
 FROM ${RUNNER_IMAGE} AS runner
