+++ pkg/cli/root_cmd.go
@@ -1,0 +1,1 @@
+package cli
